requirements.txt
pytrends
tweepy
langchain
langchain-openai
requests
python-dotenv
python-telegram-bot
asyncio